+++ inst/tinytest/test-length.R
@@ -1,0 +1,3 @@
+source("setup.R")
+
+expect_equal(length(bed), length(raw))