+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(BEDMatrix)
-
-test_check("BEDMatrix")