+++ tests/testthat/test-length.R
@@ -1,3 +1,0 @@
-test_that("length", {
-    expect_equal(length(bed), length(raw))
-})