+++ NEWS.md
@@ -1,6 +1,6 @@
 # BEDMatrix 2.0.2.9000
 
-
+- Replace testthat with tinytest.
 
 
 # BEDMatrix 2.0.2